+++ R/expr__expr.R
@@ -591,6 +591,9 @@ construct_ProtoExprArray = function(...) {
 
   # if args not named load in Expr and string
   if (is.null(arg_names)) {
+    if (length(args) == 1 && is.list(args)) {
+      args = unlist(args)
+    }
     for (i in args) {
       # if (is_string(i)) {
       #   pra$push_back_str(i)
@@ -1480,7 +1483,8 @@ Expr_mode = "use_extendr_wrapper"
 #' @description Sort this column. In projection/ selection context the whole column is sorted.
 #' If used in a groupby context, the groups are sorted.
 #' @keywords Expr
-#' @param reverse bool default FALSE, reverses sort
+#' @param descending Sort in descending order. When sorting by multiple columns,
+#' can be specified per column by passing a sequence of booleans.
 #' @param nulls_last bool, default FALSE, place Nulls last
 #' @return Expr
 #' @aliases sort
@@ -1492,8 +1496,8 @@ Expr_mode = "use_extendr_wrapper"
 #' pl$DataFrame(list(
 #'   a = c(6, 1, 0, NA, Inf, NaN)
 #' ))$select(pl$col("a")$sort())
-Expr_sort = function(reverse = FALSE, nulls_last = FALSE) { # param reverse named descending on rust side
-  .pr$Expr$sort(self, reverse, nulls_last)
+Expr_sort = function(descending = FALSE, nulls_last = FALSE) { # param reverse named descending on rust side
+  .pr$Expr$sort(self, descending, nulls_last)
 }
 
 
@@ -1544,7 +1548,8 @@ Expr_bottom_k = function(k) {
 #' @description Get the index values that would sort this column.
 #' If 'reverse=True` the smallest elements will be given.
 #' @keywords Expr
-#' @param reverse bool default FALSE, reverses sort
+#' @param descending Sort in descending order. When sorting by multiple columns,
+#' can be specified per column by passing a sequence of booleans.
 #' @param nulls_last bool, default FALSE, place Nulls last
 #' @return Expr
 #' @aliases arg_sort
@@ -1556,8 +1561,8 @@ Expr_bottom_k = function(k) {
 #' pl$DataFrame(list(
 #'   a = c(6, 1, 0, NA, Inf, NaN)
 #' ))$select(pl$col("a")$arg_sort())
-Expr_arg_sort = function(reverse = FALSE, nulls_last = FALSE) { # param reverse named descending on rust side
-  .pr$Expr$arg_sort(self, reverse, nulls_last)
+Expr_arg_sort = function(descending = FALSE, nulls_last = FALSE) { # param reverse named descending on rust side
+  .pr$Expr$arg_sort(self, descending, nulls_last)
 }
 
 
@@ -1622,7 +1627,8 @@ Expr_search_sorted = function(element) {
 #' sort column by order of others
 #' @description Sort this column by the ordering of another column, or multiple other columns.
 #' @param by one expression or list expressions and/or strings(interpreted as column names)
-#' @param reverse single bool to boolean vector, any is_TRUE will give reverse sorting of that column
+#' @param descending Sort in descending order. When sorting by multiple columns,
+#' can be specified per column by passing a sequence of booleans.
 #' @return Expr
 #' @keywords Expr
 #' @aliases sort_by
@@ -1647,13 +1653,13 @@ Expr_search_sorted = function(element) {
 #'
 #' # by two columns/expressions
 #' df$select(
-#'   pl$col("group")$sort_by(list("value2", pl$col("value1")), reverse = c(TRUE, FALSE))
+#'   pl$col("group")$sort_by(list("value2", pl$col("value1")), descending = c(TRUE, FALSE))
 #' )
 #'
 #'
 #' # by some expression
 #' df$select(
-#'   pl$col("group")$sort_by(pl$col("value1")$sort(reverse = TRUE))
+#'   pl$col("group")$sort_by(pl$col("value1")$sort(descending = TRUE))
 #' )
 #'
 #' # quite similar usecase as R function `order()`
@@ -1674,8 +1680,8 @@ Expr_search_sorted = function(element) {
 #'     pl$col("ab")$sort_by("v3")$alias("ab3"),
 #'     pl$col("ab")$sort_by("v2")$alias("ab2"),
 #'     pl$col("ab")$sort_by("v1")$alias("ab1"),
-#'     pl$col("ab")$sort_by(list("v3", pl$col("v1")), reverse = c(FALSE, TRUE))$alias("ab13FT"),
-#'     pl$col("ab")$sort_by(list("v3", pl$col("v1")), reverse = TRUE)$alias("ab13T")
+#'     pl$col("ab")$sort_by(list("v3", pl$col("v1")), descending = c(FALSE, TRUE))$alias("ab13FT"),
+#'     pl$col("ab")$sort_by(list("v3", pl$col("v1")), descending = TRUE)$alias("ab13T")
 #'   )$to_list(),
 #'   list(
 #'     ab4 = l$ab[order(l$v4)],
@@ -1686,11 +1692,11 @@ Expr_search_sorted = function(element) {
 #'     ab13T = l$ab[order(l$v3, l$v1, decreasing = TRUE)]
 #'   )
 #' )
-Expr_sort_by = function(by, reverse = FALSE) {
+Expr_sort_by = function(by, descending = FALSE) {
   .pr$Expr$sort_by(
     self,
     wrap_elist_result(by, str_to_lit = FALSE),
-    result(reverse)
+    result(descending)
   ) |> unwrap("in $sort_by:")
 }
 
@@ -2127,6 +2133,15 @@ Expr_last = "use_extendr_wrapper"
 #'   b = c("+", "-", "+", "-", "+")
 #' )$select(
 #'   pl$col("val")$count()$over("a", "b")
+#' )
+#'
+#' over_vars = c("a", "b")
+#' pl$DataFrame(
+#'   val = 1:5,
+#'   a = c("+", "+", "-", "-", "+"),
+#'   b = c("+", "-", "+", "-", "+")
+#' )$select(
+#'   pl$col("val")$count()$over(over_vars)
 #' )
 Expr_over = function(...) {
   # combine arguments in proto expression array
@@ -3238,7 +3253,7 @@ Expr_argsort = Expr_arg_sort
 #' - 'random' : Like 'ordinal', but the rank for ties is not dependent
 #' on the order that the values occur in the Series.
 #'
-#' @param reverse bool, reverse the operation
+#' @param descending Rank in descending order.
 #' @return  Expr
 #' @aliases rank
 #' @keywords Expr
@@ -3250,8 +3265,8 @@ Expr_argsort = Expr_arg_sort
 #' #  The 'ordinal' method:
 #' df = pl$DataFrame(list(a = c(3, 6, 1, 1, 6)))
 #' df$select(pl$col("a")$rank("ordinal"))
-Expr_rank = function(method = "average", reverse = FALSE) {
-  unwrap(.pr$Expr$rank(self, method, reverse))
+Expr_rank = function(method = "average", descending = FALSE) {
+  unwrap(.pr$Expr$rank(self, method, descending))
 }
 
 
@@ -4078,7 +4093,7 @@ Expr_cumulative_eval = function(expr, min_periods = 1L, parallel = FALSE) {
 #' Set_sorted
 #' @description  Flags the expression as 'sorted'.
 #* Enables downstream code to user fast paths for sorted arrays.
-#' @param reverse bool if TRUE Descending else Ascending
+#' @param descending Sort the columns in descending order.
 #' @keywords Expr
 #' @return Expr
 #' @aliases set_sorted
@@ -4090,9 +4105,9 @@ Expr_cumulative_eval = function(expr, min_periods = 1L, parallel = FALSE) {
 #' # incorrect use, flag somthing as not sorted ascendingly
 #' s2 = pl$select(pl$lit(c(1, 3, 2, 4))$set_sorted()$alias("a"))$get_column("a")
 #' s2$sort() # sorting skipped, although not actually sorted
-Expr_set_sorted = function(reverse = FALSE) {
+Expr_set_sorted = function(descending = FALSE) {
   self$map(\(s) {
-    .pr$Series$set_sorted_mut(s, reverse) # use private to bypass mut protection
+    .pr$Series$set_sorted_mut(s, descending) # use private to bypass mut protection
     s
   })
 }